pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "dreamfast"
version = "0.1.0"
description = "DreamFast: prompt → fast Blender EEVEE preview via primitives"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "", email = "" }]
dependencies = [
  "openai>=1.40.0",
  "pydantic>=2.6.1",
  "python-dotenv>=1.0.1",
  "colorama>=0.4.6",
]

# CLI entry points removed for now; use `python dreamfast.py` to start the REPL.

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
